refactor: move update especific constants to updates module
