+++ cyberdrop_dl/main.py
@@ -22,9 +22,9 @@ from cyberdrop_dl.scraper.scraper import ScrapeMapper
 from cyberdrop_dl.ui.program_ui import ProgramUI
 from cyberdrop_dl.utils import constants
 from cyberdrop_dl.utils.apprise import send_apprise_notifications
-from cyberdrop_dl.utils.check_updates import check_latest_pypi
 from cyberdrop_dl.utils.logger import RedactedConsole, log, log_spacer, log_with_color
 from cyberdrop_dl.utils.sorting import Sorter
+from cyberdrop_dl.utils.updates import check_latest_pypi
 from cyberdrop_dl.utils.utilities import check_partials_and_empty_folders, send_webhook_message
 from cyberdrop_dl.utils.yaml import handle_validation_error
