+++ cyberdrop_dl/utils/updates.py
@@ -8,15 +8,24 @@ from requests import request
 from rich.text import Text
 
 from cyberdrop_dl import __version__ as current_version
-from cyberdrop_dl.utils import constants
 from cyberdrop_dl.utils.logger import log_with_color
+
+PYPI_JSON_URL = "https://pypi.org/pypi/cyberdrop-dl-patched/json"
+PRELEASE_VERSION_PATTERN = r"(\d+)\.(\d+)\.(\d+)(?:\.([a-z]+)\d+|([a-z]+)\d+)"
+PRERELEASE_TAGS = {
+    "dev": "Development",
+    "pre": "Pre-Release",
+    "rc": "Release Candidate",
+    "a": "Alpha",
+    "b": "Beta",
+}
 
 
 def check_latest_pypi(log_to_console: bool = True, call_from_ui: bool = False) -> tuple[str, str]:
     """Checks if the current version is the latest version."""
 
     try:
-        with request("GET", constants.PYPI_JSON_URL, timeout=30) as response:
+        with request("GET", PYPI_JSON_URL, timeout=30) as response:
             contents = response.content
     except Exception:
         color = "bold_red"
@@ -56,10 +65,10 @@ def check_latest_pypi(log_to_console: bool = True, call_from_ui: bool = False) -
 
 
 def check_prelease_version(releases: list[str]) -> tuple[bool, str, Text]:
-    match = re.match(constants.PRELEASE_VERSION_PATTERN, current_version)
+    match = re.match(PRELEASE_VERSION_PATTERN, current_version)
     latest_testing_version = ""
     message = Text("")
-    running_prerelease = next((tag for tag in constants.PRERELEASE_TAGS if tag in current_version), False)
+    running_prerelease = next((tag for tag in PRERELEASE_TAGS if tag in current_version), False)
 
     if running_prerelease and match:
         major_version, minor_version, patch_version, dot_tag, no_dot_tag = match.groups()
@@ -67,7 +76,7 @@ def check_prelease_version(releases: list[str]) -> tuple[bool, str, Text]:
         regex_str = rf"{major_version}\.{minor_version}\.{patch_version}(\.{test_tag}\d+|{test_tag}\d+)"
         rough_matches = [release for release in releases if re.match(regex_str, release)]
         latest_testing_version = max(rough_matches, key=lambda x: int(re.search(r"(\d+)$", x).group()))  # type: ignore
-        ui_tag = constants.PRERELEASE_TAGS.get(test_tag, "Testing").lower()
+        ui_tag = PRERELEASE_TAGS.get(test_tag, "Testing").lower()
 
         if current_version != latest_testing_version:
             message = f"A new {ui_tag} version of Cyberdrop-DL is available: "
